requirements.txt
alembic==1.12.0
amqp==5.1.1
billiard==4.1.0
blinker==1.6.3
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flower==2.0.1
greenlet==3.0.0
humanize==4.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
SQLAlchemy==2.0.22
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
win32-setctime==1.1.0
